making atomic formulae generic over their term type introducing Term trait and removing TermBased enhancing Formula trait
