+++ razor-fol/src/transform/dnf.rs
@@ -4,14 +4,14 @@ converting [`SNF`] to [`DNF`].
 [`SNF`]: crate::transform::SNF
  */
 
-use super::{TermBased, SNF};
-use crate::syntax::{formula::*, Complex, Error, Sig, FOF, V};
+use super::SNF;
+use crate::syntax::{formula::*, term::Complex, Error, Sig, FOF, V};
 use itertools::Itertools;
 use std::ops::Deref;
 
 /// Is a [`DNF`] clause, representing conjunction of zero or more [`Literal`]s.
 #[derive(Clone, PartialEq, Eq, Hash, Debug)]
-pub struct Clause(Vec<Literal>);
+pub struct Clause(Vec<Literal<Complex>>);
 
 impl Clause {
     /// Consumes the receiver clause and `other` and returns a clause containing
@@ -23,26 +23,26 @@ impl Clause {
     }
 
     /// Returns the literals of the receiver clause.
-    pub fn literals(&self) -> &[Literal] {
+    pub fn literals(&self) -> &[Literal<Complex>] {
         &self.0
     }
 }
 
 impl Deref for Clause {
-    type Target = [Literal];
+    type Target = [Literal<Complex>];
 
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
 
-impl From<Literal> for Clause {
-    fn from(value: Literal) -> Self {
+impl From<Literal<Complex>> for Clause {
+    fn from(value: Literal<Complex>) -> Self {
         Clause(vec![value])
     }
 }
 
-impl<I: IntoIterator<Item = Literal>> From<I> for Clause {
+impl<I: IntoIterator<Item = Literal<Complex>>> From<I> for Clause {
     fn from(value: I) -> Self {
         Clause(value.into_iter().collect())
     }
@@ -50,11 +50,21 @@ impl<I: IntoIterator<Item = Literal>> From<I> for Clause {
 
 impl Default for Clause {
     fn default() -> Self {
-        Clause::from(Vec::<Literal>::new())
+        Clause::from(Vec::<Literal<Complex>>::new())
     }
 }
 
-impl TermBased for Clause {
+impl Formula for Clause {
+    type Term = Complex;
+
+    fn signature(&self) -> Result<Sig, Error> {
+        Sig::new_from_signatures(
+            self.iter()
+                .map(|l| l.signature())
+                .collect::<Result<Vec<_>, _>>()?,
+        )
+    }
+
     fn free_vars(&self) -> Vec<&V> {
         let mut vs = Vec::new();
         for lit in &self.0 {
@@ -65,16 +75,6 @@ impl TermBased for Clause {
 
     fn transform(&self, f: &impl Fn(&Complex) -> Complex) -> Self {
         Clause(self.0.iter().map(|lit| lit.transform(f)).collect())
-    }
-}
-
-impl Formula for Clause {
-    fn signature(&self) -> Result<Sig, Error> {
-        Sig::new_from_signatures(
-            self.iter()
-                .map(|l| l.signature())
-                .collect::<Result<Vec<_>, _>>()?,
-        )
     }
 }
 
@@ -156,6 +156,8 @@ impl From<&SNF> for DNF {
 }
 
 impl Formula for DNF {
+    type Term = Complex;
+
     fn signature(&self) -> Result<Sig, Error> {
         Sig::new_from_signatures(
             self.iter()
@@ -163,9 +165,7 @@ impl Formula for DNF {
                 .collect::<Result<Vec<_>, _>>()?,
         )
     }
-}
 
-impl TermBased for DNF {
     fn free_vars(&self) -> Vec<&V> {
         let mut vs = Vec::new();
         for clause in &self.0 {
@@ -225,11 +225,11 @@ fn dnf(formula: FOF) -> DNF {
         FOF::Equals(this) => Clause::from(Literal::from(this)).into(),
         FOF::Not(this) => match this.formula {
             FOF::Atom(atom) => {
-                let lit: Literal = Not { formula: atom }.into();
+                let lit: Literal<_> = Not { formula: atom }.into();
                 Clause::from(lit).into()
             }
             FOF::Equals(eq) => {
-                let lit: Literal = Not { formula: eq }.into();
+                let lit: Literal<_> = Not { formula: eq }.into();
                 Clause::from(lit).into()
             }
             _ => unreachable!(), // `formula` is in NNF
